Refactor TripDAO to enable removing its Singleton dependency
Sandro added an instance method to the TripDAO class called tripsBy().
This makes it possible for us to inject this class into the code we
have been refactoring. Because we did not remove the original static
method, no other code that depended on it will break. The new method
just passes through to the static method. As you are able to gradually
refactor all the other users of the TripDAO class, you can eventually
eiminate the original stat method findTripsByUser(). (About 1 hr mark in
 the video.)